FS Rework + mount syscall
Some major changes to how filesystems were setup, making the
system as a whole a lot more "Unix-like"

Each process not has a root directory, and the open (and mount) syscall
takes an entire path instead of a single file and parent.

To make this work there's now a layer of indirection from "struct
fs_node" called "fs_path", which is more or less analogous to Linux
"struct dentry" (though currently no caching is done unless the path is
in use). This make the mount syscall possible, because "fs_path" nodes
can have children in other "fs_mount" even while "fs_node" cannot.

Because we can't access mounts like "char:COM0" anymore, we also need to
introduce the "sys" fs_type which can be used to reference system mounts
like the "chardev" mount to access devices.

+++ include/kanawha/uapi/process.h
@@ -3,6 +3,6 @@
 
 #include <kanawha/uapi/stdint.h>
 
-typedef uintptr_t pid_t;
+typedef long pid_t;
 
 #endif
